test: Get the tests to run without the database connection
We need to set the ID of the Event and Attendee for when the save method is called. As we are using package by Layer, this means we had to add setters to the Event and Attendee classes that are public. The only way to solve this, is using a different packaging system, or creating our own dummy of the event class that does have a setter for the ID and make the ID protected instead of private.

We also used the Answer technique for mocking in our test. More info: https://semaphoreci.com/community/tutorials/stubbing-and-mocking-with-mockito-2-and-junit