Add correct Test for the current UseCase (#11)
* test: Bypass external methods from the repository like save and getAttendeeIfExists with Mockito and a mocked list of attendees.

* fix: Change Event Postgres service to show a change of state in the event after saving the new Event

* fix: Implement Comparable interface for Attendee with comparator on the PersonalCode to show when a field is Unique or how to sort.

* test: Get the tests to run without the database connection

We need to set the ID of the Event and Attendee for when the save method is called. As we are using package by Layer, this means we had to add setters to the Event and Attendee classes that are public. The only way to solve this, is using a different packaging system, or creating our own dummy of the event class that does have a setter for the ID and make the ID protected instead of private.

We also used the Answer technique for mocking in our test. More info: https://semaphoreci.com/community/tutorials/stubbing-and-mocking-with-mockito-2-and-junit

* fix: Make sure the tests don't take milliseconds into account.

* test: Test the comparison and equals methods of the Attendee class. This follows special rules for equality and comparison, entirely based on the values of the PersonalCode field.

* test: Use the doAnswer method instead of simple Mocking, as we need to return specific values that keep the flow of the UseCase method running. This is easier than a QueueMock and represents an in memory Database.

This way of working is prone to errors and is rather complicated. This can be simplified by using a better packaging system so we can easily create the needed generators or mocks to invoke with the correct access levels.

* chore: Update the todo list of the tests

---------

Co-authored-by: Nick Bauters (7dbb) <nick.bauters@colruytgroup.com>